+++ src/ui/viewer-styles/viewer-styles-options.ts
@@ -223,7 +223,6 @@ export function getOptionsStyles(): string {
     padding: 6px 12px;
     cursor: pointer;
     border-radius: 3px;
-    width: 100%;
     text-align: left;
 }
 .options-integrations-btn:hover {